Refactored code to account for SelectAsync Changes
